updated build procedure for qnx6
svn path=/trunk/yarp/; revision=2036

+++ src/ACE_wrappers/build.pl
@@ -10,7 +10,7 @@
 #		  --distribution <PATH> is the path where ACE was unpacked
 #		  --os <OS> is the operating system you're compiling for
 #
-# $Id: build.pl,v 1.4 2004-08-11 09:14:18 gmetta Exp $
+# $Id: build.pl,v 1.5 2004-10-14 20:30:11 babybot Exp $
 #
 # This script can be (at least in theory) configured to
 # do some useful thing in Linux and/or Qnx too. It's definitely
@@ -29,9 +29,6 @@ use File::Copy qw(copy);
 use strict;
 
 print "Entering build process for ACE libraries...\n";
-print "I'm trying to guess the project file name\n";
-#print "Assuming you've unpacked version 5.4.1, it might not work on other releases\n";
-#print "since Windows project names tend not to be uniform across releases\n";
 
 my $ver;
 my $uname;
@@ -39,9 +36,10 @@ my $yarp_root;
 
 chomp ($ver = `ver`);
 chomp ($uname = `uname`);
-if (index ($ver, "Windows") < 0 && index ($uname, "CYGWIN") < 0)
+if (index ($ver, "Windows") < 0 && index ($uname, "CYGWIN") < 0
+	&& index ($uname, "QNX") < 0)
 {
-	die "This script is specific to Windows 2000/XP or Cygwin\nCompilation will always happen for winnt\n";
+	die "This script is specific to Windows 2000/XP, Cygwin, or Qnx version 6\n";
 }
 
 $yarp_root = $ENV{'YARP_ROOT'};
@@ -49,6 +47,8 @@ if (!defined($yarp_root))
 {
 	die "\$YARP_ROOT environment variable must be defined!\n";
 }
+
+#require "$yarp_root/conf/configure.template.pl" or die "Can't find template file $yarp_root/conf/configure.template.pl\n";
 
 my $current_dir = getcwd;
 
@@ -61,19 +61,19 @@ my $os = undef;
 
 GetOptions ('debug' => \$debug,
             'release' => \$release,
-			'clean' => \$clean,
-			'install' => \$install,
-			'distribution=s' => \$distribution,
-			'os=s' => \$os );
+   	    'clean' => \$clean,
+	    'install' => \$install,
+	    'distribution=s' => \$distribution,
+	    'os=s' => \$os );
 
 unless (defined $distribution)
 {
 	die "This script requires the parameter --distribution <path>\n";
 }
 
-if ($os ne "winnt")
+if ($os ne "winnt" && $os ne "qnx6")
 {
-	die "This script is not yet tuned for OSes apart \"winnt\"\n";
+	die "This script is not yet tuned for OSes apart \"winnt\" and \"qnx6\"\n";
 }
 
 #
@@ -83,56 +83,13 @@ if ($os ne "winnt")
 #
 select STDERR;
 
-#
-# since ACE might have different project names I'm searching
-# on all projects in the ace directory.
-#
 my $project_name = '';
 my $file_name = '';
-foreach my $file (glob "$distribution/ace/*.dsp")
-{
-	if (open FILE, "$file") 
-	{
-		my @matching_lines = grep { /# Microsoft Developer Studio/ } <FILE>;
-		foreach my $line (@matching_lines)
-		{
-			if ($line =~ /Name=\"([\w \s]+)\" - /)
-			{
-				if ($project_name ne '')
-				{
-					my $candidate = $1;
-					if ($candidate =~ /dll/i)
-					{
-						$project_name = $candidate;
-						if ($file =~ /.*\/([\w \s]+.dsp)/)
-						{
-							$file_name = $1;
-						}
-					}
-				}
-				else
-				{
-					$project_name = $1;
-					if ($file =~ /.*\/([\w \s]+.dsp)/)
-					{
-						$file_name = $1;
-					}
-				}
-			}
-		}
 
-		close FILE;
-	}
-	else
-	{
-		warn "Cannot open $file: $!\n";
-	}
-}
-
-print "Using $file_name for project $project_name\n\n";
-if ($project_name eq '')
+if ($os eq "winnt")
 {
-	die "Can't find any suitable project name to compile ACE, sorry...\n";
+	print "I'm trying to guess the project file name\n";
+	search_project();
 }
 
 if ($clean)
@@ -140,9 +97,16 @@ if ($clean)
 	print "\nCleaning...\n";
 	chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
 
-	call_msdev_and_print ("Debug", "CLEAN", $project_name, $file_name);
-	call_msdev_and_print ("Release", "CLEAN", $project_name, $file_name);
-	
+	if ($os eq "winnt")
+	{
+		call_msdev_and_print ("Debug", "CLEAN", $project_name, $file_name);
+		call_msdev_and_print ("Release", "CLEAN", $project_name, $file_name);
+	}
+	elsif ($os eq "qnx6")
+	{
+		call_make_and_print ('', 'clean');
+	}
+
 	print "\n";
 	chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
 }
@@ -151,26 +115,52 @@ if ($debug)
 {
 	print "\nCompiling debug\n";
 
-	copy ("$yarp_root/include/winnt/ace/config.h", "$distribution/ace/") or die "Can't copy config.h file\n"; 
-	chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
+	if ($os eq "winnt")
+	{
+		copy ("$yarp_root/include/winnt/ace/config.h", "$distribution/ace/") or die "Can't copy config.h file\n"; 
+		chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
 
-	call_msdev_and_print ("Debug", "BUILD", $project_name, $file_name);
+		call_msdev_and_print ("Debug", "BUILD", $project_name, $file_name);
 	
-	unlink "config.h";
-	chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
+		unlink "config.h";
+		chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
+	}
+	elsif ($os eq "qnx6")
+	{
+		chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
+		symlink ("$yarp_root/include/qnx6/ace/platform_qnx_rtp_gcc.GNU", "$distribution/include/makeinclude/platform_macros.GNU");
+		symlink ("$yarp_root/include/qnx6/ace/config-qnx-rtp-62x.h", "$distribution/ace/config.h");
+
+		call_make_and_print ('', 'debug=1 optimize=0');
+	
+		chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
+	}
 }
 
 if ($release)
 {
 	print "\nCompiling release\n";
 
-	copy ("$yarp_root/include/winnt/ace/config.h", "$distribution/ace/") or die "Can't copy config.h file\n"; 
-	chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
+	if ($os eq "winnt")
+	{
+		copy ("$yarp_root/include/winnt/ace/config.h", "$distribution/ace/") or die "Can't copy config.h file\n"; 
+		chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
 
-	call_msdev_and_print ("Release", "BUILD", $project_name, $file_name);
+		call_msdev_and_print ("Release", "BUILD", $project_name, $file_name);
 	
-	unlink "config.h";
-	chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
+		unlink "config.h";
+		chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
+	}
+	elsif ($os eq "qnx6")
+	{
+		chdir "$distribution/ace" or die "Cannot chdir to $distribution/ace: $!";
+		symlink ("$yarp_root/include/qnx6/ace/platform_qnx_rtp_gcc.GNU", "$distribution/include/makeinclude/platform_macros.GNU");
+		symlink ("$yarp_root/include/qnx6/ace/config-qnx-rtp-62x.h", "$distribution/ace/config.h");
+
+		call_make_and_print ('', 'optimize=1 debug=0');
+	
+		chdir "$current_dir" or die "Cannot chdir to $current_dir: $!";
+	}
 }
 
 if ($install)
@@ -207,6 +197,75 @@ if ($install)
 }
 
 select STDOUT;
+
+#
+# since ACE might have different project names I'm searching
+# on all projects in the ace directory.
+#
+sub search_project
+{
+	foreach my $file (glob "$distribution/ace/*.dsp")
+	{
+		if (open FILE, "$file") 
+		{
+			my @matching_lines = grep { /# Microsoft Developer Studio/ } <FILE>;
+			foreach my $line (@matching_lines)
+			{
+				if ($line =~ /Name=\"([\w \s]+)\" - /)
+				{
+					if ($project_name ne '')
+					{
+						my $candidate = $1;
+						if ($candidate =~ /dll/i)
+						{
+							$project_name = $candidate;
+							if ($file =~ /.*\/([\w \s]+.dsp)/)
+							{
+								$file_name = $1;
+							}
+						}
+					}
+					else
+					{
+						$project_name = $1;
+						if ($file =~ /.*\/([\w \s]+.dsp)/)
+						{
+							$file_name = $1;
+						}
+					}
+				}
+			}
+	
+			close FILE;
+		}
+		else
+		{
+			warn "Cannot open $file: $!\n";
+		}
+	}
+	
+	print "Using $file_name for project $project_name\n\n";
+	if ($project_name eq '')
+	{
+		die "Can't find any suitable project name to compile ACE, sorry...\n";
+	}
+}
+
+#
+#
+#
+sub call_make_and_print
+{
+        my ($project, $operation) = @_;
+
+        open MK, "make ACE $operation"."|";
+        while (<MK>)
+        {
+                print;
+        }
+        close MK;
+}
+
 
 #
 #
